chore: move models and other weird impls
