+++ crates/storage/db-api/src/models/accounts.rs
@@ -2,9 +2,12 @@
 
 use std::ops::{Range, RangeInclusive};
 
-use crate::{impl_fixed_arbitrary, DatabaseError};
+use crate::{
+    impl_fixed_arbitrary,
+    table::{Decode, Encode},
+    DatabaseError,
+};
 use reth_codecs::{derive_arbitrary, Compact};
-use reth_db_api::table::{Decode, Encode};
 use reth_primitives::{Account, Address, BlockNumber, Buf, StorageKey};
 use serde::{Deserialize, Serialize};
