+++ crates/storage/db-api/src/models/storage_sharded_key.rs
@@ -1,8 +1,9 @@
 //! Storage sharded key
-
-use crate::DatabaseError;
+use crate::{
+    table::{Decode, Encode},
+    DatabaseError,
+};
 use derive_more::AsRef;
-use reth_db_api::table::{Decode, Encode};
 use reth_primitives::{Address, BlockNumber, B256};
 use serde::{Deserialize, Serialize};
