+++ crates/storage/db/src/tables/codecs/mod.rs
@@ -1,8 +1,3 @@
 //! Integrates different codecs into `table::Encode` and `table::Decode`.
 
-mod compact;
-pub use compact::CompactU256;
-
 pub mod fuzz;
-
-mod scale;
