--- crates/storage/db/src/tables/models/mod.rs
@@ -1,176 +1,0 @@
-//! Implements data structures specific to the database
-
-use crate::DatabaseError;
-use reth_codecs::Compact;
-use reth_db_api::table::{Decode, Encode};
-use reth_primitives::{
-    trie::{StoredNibbles, StoredNibblesSubKey},
-    Address, PruneSegment, B256,
-};
-
-pub mod accounts;
-pub mod blocks;
-pub mod client_version;
-pub mod integer_list;
-pub mod sharded_key;
-pub mod storage_sharded_key;
-
-pub use accounts::*;
-pub use blocks::*;
-pub use sharded_key::ShardedKey;
-
-use self::client_version::ClientVersion;
-
-/// Macro that implements [`Encode`] and [`Decode`] for uint types.
-macro_rules! impl_uints {
-    ($($name:tt),+) => {
-        $(
-            impl Encode for $name {
-                type Encoded = [u8; std::mem::size_of::<$name>()];
-
-                fn encode(self) -> Self::Encoded {
-                    self.to_be_bytes()
-                }
-            }
-
-            impl Decode for $name {
-                fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, $crate::DatabaseError> {
-                    Ok(
-                        $name::from_be_bytes(
-                            value.as_ref().try_into().map_err(|_| $crate::DatabaseError::Decode)?
-                        )
-                    )
-                }
-            }
-        )+
-    };
-}
-
-impl_uints!(u64, u32, u16, u8);
-
-impl Encode for Vec<u8> {
-    type Encoded = Self;
-
-    fn encode(self) -> Self::Encoded {
-        self
-    }
-}
-
-impl Decode for Vec<u8> {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        Ok(value.as_ref().to_vec())
-    }
-}
-
-impl Encode for Address {
-    type Encoded = [u8; 20];
-
-    fn encode(self) -> Self::Encoded {
-        self.0 .0
-    }
-}
-
-impl Decode for Address {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        Ok(Self::from_slice(value.as_ref()))
-    }
-}
-
-impl Encode for B256 {
-    type Encoded = [u8; 32];
-
-    fn encode(self) -> Self::Encoded {
-        self.0
-    }
-}
-
-impl Decode for B256 {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        Ok(Self::new(value.as_ref().try_into().map_err(|_| DatabaseError::Decode)?))
-    }
-}
-
-impl Encode for String {
-    type Encoded = Vec<u8>;
-
-    fn encode(self) -> Self::Encoded {
-        self.into_bytes()
-    }
-}
-
-impl Decode for String {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        Self::from_utf8(value.as_ref().to_vec()).map_err(|_| DatabaseError::Decode)
-    }
-}
-
-impl Encode for StoredNibbles {
-    type Encoded = Vec<u8>;
-
-    // Delegate to the Compact implementation
-    fn encode(self) -> Self::Encoded {
-        let mut buf = Vec::with_capacity(self.0.len());
-        self.to_compact(&mut buf);
-        buf
-    }
-}
-
-impl Decode for StoredNibbles {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        let buf = value.as_ref();
-        Ok(Self::from_compact(buf, buf.len()).0)
-    }
-}
-
-impl Encode for StoredNibblesSubKey {
-    type Encoded = Vec<u8>;
-
-    // Delegate to the Compact implementation
-    fn encode(self) -> Self::Encoded {
-        let mut buf = Vec::with_capacity(65);
-        self.to_compact(&mut buf);
-        buf
-    }
-}
-
-impl Decode for StoredNibblesSubKey {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        let buf = value.as_ref();
-        Ok(Self::from_compact(buf, buf.len()).0)
-    }
-}
-
-impl Encode for PruneSegment {
-    type Encoded = [u8; 1];
-
-    fn encode(self) -> Self::Encoded {
-        let mut buf = [0u8];
-        self.to_compact(&mut buf.as_mut());
-        buf
-    }
-}
-
-impl Decode for PruneSegment {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        let buf = value.as_ref();
-        Ok(Self::from_compact(buf, buf.len()).0)
-    }
-}
-
-impl Encode for ClientVersion {
-    type Encoded = Vec<u8>;
-
-    // Delegate to the Compact implementation
-    fn encode(self) -> Self::Encoded {
-        let mut buf = vec![];
-        self.to_compact(&mut buf);
-        buf
-    }
-}
-
-impl Decode for ClientVersion {
-    fn decode<B: AsRef<[u8]>>(value: B) -> Result<Self, DatabaseError> {
-        let buf = value.as_ref();
-        Ok(Self::from_compact(buf, buf.len()).0)
-    }
-}
