+++ crates/storage/db/src/tables/raw.rs
@@ -13,7 +13,7 @@ pub struct RawTable<T: Table> {
 }
 
 impl<T: Table> Table for RawTable<T> {
-    const TABLE: crate::Tables = T::TABLE;
+    const NAME: &'static str = T::NAME;
 
     type Key = RawKey<T::Key>;
     type Value = RawValue<T::Value>;
@@ -27,7 +27,7 @@ pub struct RawDupSort<T: DupSort> {
 }
 
 impl<T: DupSort> Table for RawDupSort<T> {
-    const TABLE: crate::Tables = T::TABLE;
+    const NAME: &'static str = T::NAME;
 
     type Key = RawKey<T::Key>;
     type Value = RawValue<T::Value>;
